models/football_result_prediction/.history/blogapp/requirements_20250106232417.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
diff-match-patch==20241021
Django==5.1.4
django-import-export==4.3.3
executing==2.0.1
h11==0.14.0
idna==3.10
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.2
pywin32==306
pyzmq==25.1.2
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stack-data==0.6.3
tablib==3.7.0
tornado==6.4
traitlets==5.14.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
